api_gateway/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
python-dotenv
requests
